flask_ocr/requirements.txt
Flask
pillow
pytesseract
transformers
sentencepiece
protobuf